Use longs instead of integer
Based on the uv since version 5.x the ids are now longs.

+++ model-events/src/main/java/de/focus_shift/urlaubsverwaltung/extension/api/application/ApplicationCreatedFromSickNoteEventDTO.java
@@ -17,7 +17,7 @@ public class ApplicationCreatedFromSickNoteEventDTO {
 
     @NonNull private UUID id;
 
-    @NonNull private Integer sourceId;
+    @NonNull private Long sourceId;
 
     @NonNull private Instant createdAt;
